==> exp2/analysis_scripts/0b_ProcessData.Rmd <==
---
title: "Violations of physical and psychological expectations in the human adult brain - Prepare Experiment 2 data for analysis"
author: "Shari Liu"
date: '2023-08-02'
output: html_document
---

```{r setup}
knitr::opts_chunk$set(echo = TRUE)
library(pacman)
pacman::p_load(tidyverse,
               data.table,
               janitor,
               conflicted,
               here,
               reticulate,
               lme4,
               lmerTest,
               EMAtools,
               patchwork,
               cowplot)

conflict_prefer("select", "dplyr")
conflict_prefer("filter", "dplyr")
conflict_prefer("rename", "dplyr")
conflict_prefer("arrange", "dplyr")
conflict_prefer("summarise", "dplyr")
conflict_prefer("mutate", "dplyr")

conflict_prefer("lmer", "lmerTest")
conflict_prefer("here", "here")

source(here("helper_funs.R"))
source_python(here("pickle_reader.py"))

sessionInfo()

```

# Overview
This script processes Study 2 data for univariate and multivariate analysis. All instances of `write_rds` `write_csv` and similar have been commented out to avoid writing over existing files. Uncomment those statements in order to reproduce these steps and save (new) outputs.

# Main ROI data setup
```{r}
# read in data, and information about ROIs
factors <-
  c("ROI_name", "event", "top_n_voxels", "domain", "event", "extracted_copes_main_condition", "extracted_run_number")

study2.data0 <-
  read_csv(here("input_data/ROI_analysis_outputs_Apr06_2023_convertVectors.csv")) %>%
  filter(!is.na(extracted_run_number)) %>%
  separate(
    col = "extracted_copes_main_condition",
    into = c("domain", "event") ,
    # into = c("domain", "task", "event") ,
    sep = "_",
    remove = FALSE
  ) %>%
  rename(meanbeta = mean_topVoxels_main_condition_copes,
         top_n_voxels = top_voxel_selection_method) %>%
  mutate_at(factors, as.factor)

region_info <- read.csv(here("input_data/manyregions_info.csv")) %>%
  mutate(ROI_name = paste0(ROI_name, "_MNI152Lin2009cAsym_res-02", sep = ""))

subject_info <- read.csv(here("input_data/exp2_study_subjects.csv"))

table(subject_info$gender)
table(subject_info$handed)
summary(subject_info)
study2.data.prelim <- full_join(study2.data0, region_info) %>%
mutate(LR = case_when(str_detect(ROI_name_final, "bilateral") ~ "bilateral",
                      TRUE ~ str_sub(ROI_name_final,-1,-1)))

study2.data <- full_join(study2.data.prelim, subject_info) 
study2.data$event <- relevel(study2.data$event, ref = "fam")
```

```{r}
# are the number of observations per region per run per subject expected? n_observations should be 9 (3 events x 3 domains)
study2.data %>%
  group_by(top_n_voxels, subjectID, ROI_name, extracted_run_number) %>%
  summarise(n_observations = n()) 

# check for duplicates
# they exist, but only for parcels that have bilateral counterparts, because the selected voxels are the same across the two
study2.data %>%
  get_dupes(vector1_topVoxels_main_condition_copes) %>%
  group_by(top_n_voxels, subjectID, ROI_name, extracted_run_number) %>%
  summarise(n_observations = n()) 
```

```{r}
study2.data.top100 <- study2.data %>%
  filter(top_n_voxels == 100)
```

```{r}
# saveRDS(study2.data.top100, here("outputs/univariate_data/study2_univariate_data_top100.Rds"))
# saveRDS(study2.data, here("outputs/univariate_data/study2_univariate_data_allregions_alltopN.Rds"))
```

```{r}
study2.data <- read_rds( here("outputs/univariate_data/study2_univariate_data_allregions_alltopN.Rds"))

```

## Summary statistics

The missing values in the table below is due to sub-SAXNES2s022, who has runs 1, 2 and 4 from the VOE task included in the analysis. Thus, their folds are (1) excluding run 1, (2) excluding run 2, and (3) excluding run 4. 

```{r}
# summary_task <-
#   summarySEwithin(
#     data = study2.data,
#     measurevar = "meanbeta",
#     withinvars = c("top_n_voxels", "ROI_name_final", "ROI_category", "extracted_run_number", "focal_region", "domain", "task", "event"),
#     idvar = "subjectID"
#   )

summary_domain <-
  summarySEwithin(
    data = study2.data,
    measurevar = "meanbeta",
    withinvars = c("top_n_voxels", "ROI_name_final", "ROI_category", "old_ROI", "LR", "extracted_run_number", "focal_region", "domain", "event"),
    idvar = "subjectID"
  )

summary_domain_runs12 <- 
  summarySEwithin(
    data = study2.data %>% filter(extracted_run_number == "run1" | extracted_run_number == "run2"),
    measurevar = "meanbeta",
    withinvars = c("top_n_voxels", "ROI_name_final", "ROI_category","old_ROI", "LR", "focal_region", "domain", "event"),
    idvar = "subjectID"
  )

summary_domain_allruns <-
  summarySEwithin(
    data = study2.data %>% filter(str_length(extracted_run_number) == 4),
    measurevar = "meanbeta",
    withinvars = c("top_n_voxels", "ROI_name_final", "ROI_category", "old_ROI", "LR", "focal_region", "domain", "event"),
    idvar = "subjectID"
  )
```

```{r}
# saveRDS(summary_domain, here("outputs", "descriptive_summaries", "study2_univariate_summary_domain.Rds"))
# saveRDS(summary_domain_runs12, here("outputs", "descriptive_summaries", "study2_univariate_summary_domain_runs12.Rds"))
# saveRDS(summary_domain_allruns, here("outputs", "descriptive_summaries", "study2_univariate_summary_domain_allruns.Rds"))
```

### Subject info
```{r}
study2.subjects <- study2.data %>%
  select(subjectID, data_sharing, gender, handed, age, fixation_position) %>%
  unique()
```

# Set up data for sensitivity analysis

```{r}
study2.data <- read_rds(here("outputs/univariate_data/study2_univariate_data_allregions_alltopN.Rds"))
```

```{r}
summary_domain_all_topvoxel_N <-
  summarySEwithin(
    data = study2.data,
    measurevar = "meanbeta",
    withinvars = c("top_n_voxels", "ROI_name_final", "ROI_category", "LR", "extracted_run_number", "focal_region", "domain", "event"),
    idvar = "subjectID"
  )

summary_domain_all_topvoxel_N_runs12 <-
  summarySEwithin(
    data = study2.data %>% filter(extracted_run_number == "run1" | extracted_run_number == "run2"),
    measurevar = "meanbeta",
    withinvars = c("top_n_voxels", "ROI_name_final", "ROI_category", "LR", "focal_region", "domain", "event"),
    idvar = "subjectID"
  )
```

```{r}
# saveRDS(summary_domain_all_topvoxel_N, here("outputs", "descriptive_summaries", "study2_univariate_summary_domain_all_topvoxel_N.Rds"))
# saveRDS(summary_domain_all_topvoxel_N_runs12, here("outputs", "descriptive_summaries", "study2_univariate_summary_domain_all_topvoxel_N.Rds"))
```

# VOE runs 3-4 ROI data setup

```{r}
# read in data, and information about ROIs
factors <-
  c("ROI_name", "event", "top_n_voxels", "domain", "event", "extracted_copes_main_condition", "extracted_run_number")

study2.VOEextract.data0 <-
  read.csv(here("input_data/ROI_analysis_VOErun34extract_outputs_Apr08_2023_convertVectors.csv")) %>%
  filter(top_voxel_selection_method == 100,
         !is.na(extracted_run_number)) %>%
  separate(
    col = "extracted_copes_main_condition",
    into = c("domain", "event") ,
    # into = c("domain", "task", "event") ,
    sep = "_",
    remove = FALSE
  ) %>%
  rename(meanbeta = mean_topVoxels_main_condition_copes,
         top_n_voxels = top_voxel_selection_method) %>%
  mutate_at(factors, as.factor)

region_info <- read.csv(here("input_data/manyregions_info.csv")) %>%
  mutate(ROI_name = paste0(ROI_name, "_MNI152Lin2009cAsym_res-02", sep = ""))

subject_info <- read.csv(here("input_data/exp2_study_subjects.csv")) %>%
  filter(include == 1) %>%
  select(subjectID, fixation_position, gender, handed, age)

study2.VOEextract.data.prelim <- full_join(study2.VOEextract.data0, region_info) %>%
mutate(LR = case_when(str_detect(ROI_name_final, "bilateral") ~ "bilateral",
                      TRUE ~ str_sub(ROI_name_final,-1,-1)))

study2.VOEextract.data <- full_join(study2.VOEextract.data.prelim, subject_info) 
study2.VOEextract.data$event <- relevel(study2.VOEextract.data$event, ref = "fam")

```

```{r}
# are the number of observations per region per run per subject expected? n_observations should be 9 (3 events x 3 domains)
study2.VOEextract.data %>%
  group_by(top_n_voxels, subjectID, ROI_name, extracted_run_number) %>%
  summarise(n_observations = n()) 

# check for duplicates
study2.VOEextract.data %>%
  # filter(subjectID == "sub-SAXNES2s028") %>%
  get_dupes(vector1_topVoxels_main_condition_copes) %>%
  group_by(top_n_voxels, subjectID, ROI_name, extracted_run_number) %>%
  summarise(n_duplicates = n()) 
```

```{r}
str(study2.VOEextract.data)
```

```{r}
# saveRDS(study2.VOEextract.data, here("outputs/univariate_data/study2_univariate_data_VOEextract.Rds"))
```


## Summary statistics

```{r}
summary_domain <-
  summarySEwithin(
    data = study2.VOEextract.data,
    measurevar = "meanbeta",
    withinvars = c("top_n_voxels", "ROI_name_final", "ROI_category", "LR", "extracted_run_number", "focal_region", "domain", "event"),
    idvar = "subjectID"
  )

summary_domain_runs12 <-
  summarySEwithin(
    data = study2.VOEextract.data %>% filter(extracted_run_number == "run1" | extracted_run_number == "run2"),
    measurevar = "meanbeta",
    withinvars = c("top_n_voxels", "ROI_name_final", "ROI_category", "LR", "focal_region", "domain", "event"),
    idvar = "subjectID"
  )
```

```{r}
# saveRDS(summary_domain, here("outputs", "descriptive_summaries", "study2_univariate_summary_domain_VOEextract.Rds"))
# saveRDS(summary_domain_runs12, here("outputs", "descriptive_summaries", "study2_univariate_summary_domain_runs12_VOEextract.Rds"))

```


# VOE single betas ROI data setup

```{r}
# read in data, and information about ROIs
factors <-
  c("ROI_name", "event", "top_n_voxels", "domain", "event", "extracted_copes_main_condition", "extracted_run_number")


study2.singlebetas.data0 <-
  read.csv(here("input_data/ROI_analysis_singlebetas_outputs_Apr11_2023_convertVectors.csv")) %>%
  filter(extracted_copes_main_condition != "check") %>% # filter out attention check betas
  mutate(contrast_for_topVoxel_selection = str_remove_all(contrast_for_topVoxel_selection, "contrast-"),
         selection_contrast_task_name = str_remove_all(selection_contrast_task_name, "task-"),
         ROI_name = str_remove_all(ROI_name, "_MNI152Lin2009cAsym_res-02")) %>%
  filter(top_voxel_selection_method == 100,
         !is.na(extracted_run_number)) %>%
  separate(
    col = "extracted_copes_main_condition",
    into = c("domain", "task", "scenario", "event") ,
    sep = "_",
    remove = FALSE
  ) %>%
  rename(meanbeta = mean_topVoxels_main_condition_copes,
         top_n_voxels = top_voxel_selection_method) %>%
  mutate_at(factors, as.factor) %>%
  mutate(scenario = case_when(str_detect(scenario, "exp1") ~ "exp1",
                                                               TRUE ~ scenario)) %>%
  unite("extracted_copes_main_condition", c("domain", "task", "scenario", "event"), remove = FALSE) %>%
  unite("ROI_outputs_video_group", c("domain", "task", "scenario"), remove = FALSE) 


region_info <- read.csv(here("input_data/manyregions_info.csv")) %>%
  select(-ROI_category)

subject_info <- read.csv(here("input_data/exp2_study_subjects.csv")) %>%
  filter(include == 1) %>%
  select(subjectID, fixation_position, gender, handed, data_sharing, age)

condition_info <- 
  read.csv(here("../experiment_scripts/VOE/stimulus_schedules/exp2_paths.csv")) %>%
  filter(include_exp2 == 1)

unique(sort(condition_info$vis_statistics_video_group)) # this is the common col
unique(sort(condition_info$ROI_outputs_video_group))


# join ROI data to region data
study2.singlebetas.data.prelim <- full_join(study2.singlebetas.data0, region_info, by = c("ROI_name")) %>%
mutate(LR = case_when(str_detect(ROI_name_final, "bilateral") ~ "bilateral",
                      TRUE ~ str_sub(ROI_name_final,-1,-1)))

# join these data to subject data
study2.singlebetas.data.prelim2 <- full_join(study2.singlebetas.data.prelim, subject_info) 

# join these data to condition data <-
study2.singlebetas.data.prelim3 <- full_join(study2.singlebetas.data.prelim2, condition_info, by = c("ROI_outputs_video_group")) %>%
  select(-X) %>%
  filter(!is.na(top_n_voxels))

study2.singlebetas.data.prelim4 <- study2.singlebetas.data.prelim3 %>%
  unite("video", c("domain", "task", "scenario", "event"), remove = FALSE)

# read in visual stats data
vis_stats_info_all <- read.csv(here("outputs/visual_statistics/visual_statistics_exp2_new.csv")) %>% select(-X, -domain, -task) %>%
  filter(in_exp2 == TRUE)

vis_stats_info_unnormalized <-
  vis_stats_info_all %>%
  select(-normalized_value, -X.1) %>%
  pivot_wider(
    id_cols = c(
      "vis_statistics_video_group",
      "event",
      "event_orig",
      # "extracted_copes_main_condition",
      "in_exp1",
      "in_exp2",
      "identical_test"
    ),
    names_from = "metric",
    values_from = "value"
  )

vis_stats_info_normalized <-
  vis_stats_info_all %>%
  select(-value, -X.1) %>%
  pivot_wider(
    id_cols = c(
      "vis_statistics_video_group",
      "event",
      "event_orig",
      # "extracted_copes_main_condition",
      "in_exp1",
      "in_exp2",
      "identical_test"
    ),
    names_from = "metric",
    values_from = "normalized_value",
    names_prefix = "normalized_"
    )

vis_stats_info_tidied <- full_join(vis_stats_info_normalized, vis_stats_info_unnormalized)

unique(sort(vis_stats_info_tidied$vis_statistics_video_group)) == unique(sort(study2.singlebetas.data.prelim4$vis_statistics_video_group))

study2.singlebetas.data <- left_join(study2.singlebetas.data.prelim4, vis_stats_info_tidied, by = c("vis_statistics_video_group", "event")) %>%
  unique() %>%
  select(-n_voxels_in_brainmasked_parcel,
         -n_voxels_subject_specific_brainmasked_parcel,
         -ROI_outputs_video_group,
         -img1,
         -img2,
         -fam,
         -exp,
         -unexp) %>%
  filter(!is.na(extracted_run_number),
         !is.na(subjectID)) 

study2.singlebetas.data$task <- factor(study2.singlebetas.data$task, c("solidity", "permanence", "efficiency", "goal", "agent-solidity", "infer-constraint"))

study2.singlebetas.data$domain <- factor(study2.singlebetas.data$domain, c("physics", "psychology", "both"))

levels(study2.singlebetas.data$domain) <- c("physics", "psychology-action", "psychology-environment")
# write_rds(study2.singlebetas.data, here("outputs/univariate_data/study2_univariate_data_singlebetas.Rds"))

```

```{r}
study2.singlebetas.data <- read_rds( here("outputs/univariate_data/study2_univariate_data_singlebetas.Rds")) 

```

```{r}
# are the number of observations per region per run per subject expected? n_observations should be 48 (psych-environment) or 96 (psych-action, physics)
study2.singlebetas.counts <- study2.singlebetas.data %>% 
  group_by(top_n_voxels, subjectID, extracted_copes_main_condition, extracted_run_number) %>%
  summarise(n_observations = n()) 

range(study2.singlebetas.counts$n_observations)
```

## Summary statistics

```{r}

# per run per domain
summary_domain <-
  summarySEwithin(
    data = study2.singlebetas.data,
    measurevar = "meanbeta",
    withinvars = c("top_n_voxels", "ROI_name_final", "ROI_category", "LR", "extracted_run_number", "focal_region", "domain", "event"),
    idvar = "subjectID"
  )

# per run per task
summary_task <-
  summarySEwithin(
    data = study2.singlebetas.data,
    measurevar = "meanbeta",
    withinvars = c("top_n_voxels", "ROI_name_final", "ROI_category", "LR", "extracted_run_number", "focal_region", "domain", "task", "event"),
    idvar = "subjectID"
  )

# per run per scenario
summary_scenario <-
  summarySEwithin(
    data = study2.singlebetas.data, 
    # %>% filter(extracted_run_number == "run1" | extracted_run_number == "run2"),
    measurevar = "meanbeta",
    withinvars = c("top_n_voxels", "ROI_name_final", "ROI_category", "LR", "focal_region", "domain", "vis_statistics_video_group", "in_exp1", "task", "event"),
    idvar = "subjectID"
  )
```

```{r}
saveRDS(summary_domain, here("outputs/descriptive_summaries/study2_univariate_summary_domain_singlebetas.Rds"))
saveRDS(summary_task, here("outputs/descriptive_summaries/study2_univariate_summary_task_singlebetas.Rds"))
saveRDS(summary_scenario, here("outputs/descriptive_summaries/study2_univariate_summary_scenario_singlebetas.Rds"))

```

### Subject info
```{r}
study2.subjects <- study2.data %>%
  select(subjectID, consent_openneuro_sharing, gender, handedness, age_years, fixation_position) %>%
  unique()
```

### Size of parcels
```{r}
range(study2.data$n_voxels_in_brainmasked_parcel)
range(study2.data$n_voxels_subject_specific_brainmasked_parcel)

```


# Set up multivariate data



```{r}
study2.data <- readRDS(here("outputs/univariate_data/study2_univariate_data_allregions_alltopN.Rds")) %>%
  filter(top_n_voxels == "100")

# first we need to unpack vectors into their own columns
MVPA_vectors0 <- study2.data %>%
  select(
    top_n_voxels,
    extracted_run_number,
    focal_region,
    subjectID,
    ROI_name_final,
    ROI_category,
    contrast_for_topVoxel_selection,
    extracted_copes_main_condition,
    domain,
    # task,
    event,
    # rep,  ## only relevant for analysis that splits conditions by repetition (2 per run)
    vector1_topVoxels_main_condition_copes
  )

# only runs 1 and 2
MVPA_vectors_runs12 <- MVPA_vectors0 %>%
  filter(extracted_run_number == "run1" | extracted_run_number == "run2") 

MVPA_vectors <- data.frame()
  
cur_subset <- MVPA_vectors_runs12

for (i in 1:nrow(cur_subset)) {

  # get all relevant info associated with each cope vector
  top_n_voxels <- cur_subset$top_n_voxels[i]
  focal_region <- cur_subset$focal_region[i]
  subj <- cur_subset$subjectID[i]
  ROI <- cur_subset$ROI_name_final[i]
  ROI_category <- cur_subset$ROI_category[i]
  selection_contrast <- cur_subset$contrast_for_topVoxel_selection[i]
  domain <- cur_subset$domain[i]
  # task <- cur_subset$task[i]
  event <- cur_subset$event[i]
  extracted_run_number <- cur_subset$extracted_run_number[i]
  # rep <- cur_subset$rep[i]  ## only relevant for analysis that splits conditions by repetition (2 per run)
  cond <- cur_subset$extracted_copes_main_condition[i]

  # get the cope vector and convert into cols (Vox1-100)
  copes <- cur_subset$vector1_topVoxels_main_condition_copes[i] %>%
    as.vector() %>%
    str_split(., pattern = ", ") %>%
    flatten_chr() %>%
    str_remove_all(., patter = "]") %>%
    str_remove_all(., pattern = "\\[") %>%
    as.numeric() %>%
    t() %>%
    as.data.frame()
  
  # record actual length of copes to check for issues w extraction
  length_cope_actual <- length(copes)
  
  # combine all of that into a single row for the df
  cur.row <- cbind(top_n_voxels,
                   length_cope_actual,
                   extracted_run_number,
                   focal_region,
                   subj,
                   ROI,
                   selection_contrast,
                   domain,
                   # task,
                   event,
                   cond,
                   copes)
    
    # bind to full df
    MVPA_vectors <- rbind(MVPA_vectors, cur.row)
    # newlist <- list(MVPA_vectors, cur.row)
    # MVPA_vectors <- rbindlist(newlist, fill = TRUE)
}

# range(MVPA_vectors$length_cope_actual)

# saveRDS(MVPA_vectors, here("outputs", "multivariate_data", "study2_multivariate_data_runs12_top100.Rds"))
```


```{r}

# only LOO and single runs
MVPA_vectors_folds <- MVPA_vectors0 %>%
  filter(str_length(extracted_run_number) == 4 | str_detect(extracted_run_number, "excluderun"))

MVPA_vectors_allfolds <- data.frame()
  
cur_subset <- MVPA_vectors_folds

for (i in 1:nrow(cur_subset)) {

  # get all relevant info associated with each cope vector
  top_n_voxels <- cur_subset$top_n_voxels[i]
  focal_region <- cur_subset$focal_region[i]
  subj <- cur_subset$subjectID[i]
  ROI <- cur_subset$ROI_name_final[i]
  ROI_category <- cur_subset$ROI_category[i]
  selection_contrast <- cur_subset$contrast_for_topVoxel_selection[i]
  domain <- cur_subset$domain[i]
  # task <- cur_subset$task[i]
  event <- cur_subset$event[i]
  extracted_run_number <- cur_subset$extracted_run_number[i]
  # rep <- cur_subset$rep[i]  ## only relevant for analysis that splits conditions by repetition (2 per run)
  cond <- cur_subset$extracted_copes_main_condition[i]

  # get the cope vector and convert into cols (Vox1-100)
  copes <- cur_subset$vector1_topVoxels_main_condition_copes[i] %>%
    as.vector() %>%
    str_split(., pattern = ", ") %>%
    flatten_chr() %>%
    str_remove_all(., patter = "]") %>%
    str_remove_all(., pattern = "\\[") %>%
    as.numeric() %>%
    t() %>%
    as.data.frame()
  
  # record actual length of copes to check for issues w extraction
  length_cope_actual <- length(copes)
  
  # combine all of that into a single row for the df
  cur.row <- cbind(top_n_voxels,
                   length_cope_actual,
                   extracted_run_number,
                   focal_region,
                   subj,
                   ROI,
                   selection_contrast,
                   domain,
                   # task,
                   event,
                   cond,
                   copes)
    
    # bind to full df
    MVPA_vectors_allfolds <- rbind(MVPA_vectors_allfolds, cur.row)
    # newlist <- list(MVPA_vectors, cur.row)
    # MVPA_vectors <- rbindlist(newlist, fill = TRUE)
}

range(MVPA_vectors_allfolds$length_cope_actual)

# saveRDS(MVPA_vectors_allfolds, here("outputs/multivariate_data/study2_multivariate_data_acrossfolds_top100.Rds"))

```

## Set up distances

### Primary analysis
Here we only want to take the test events from psychology and physics from runs 1 and 2.

```{r}
MVPA_vectors_runs12 <-
  readRDS(
    here(
      "outputs",
      "multivariate_data",
      "study2_multivariate_data_runs12_top100.Rds"
    )
  )

# are the number of observations per region per subject expected? n_observations should be 9 (2 runs, 3 domains, 3 events) across 4 runs, 3 per run
MVPA_vectors_runs12 %>%
  group_by(subj, ROI, extracted_run_number) %>%
  summarise(n_observations = n()) 

table(MVPA_vectors_runs12$event)
table(MVPA_vectors_runs12$length_cope_actual)
```

```{r}
# z score vectors within voxels within runs within ROIs within subjects
MVPA_vectors_runs12_zscored0 <- MVPA_vectors_runs12 %>%
  gather(voxel, beta, V1:V100) %>%
  group_by(subj, ROI, extracted_run_number, voxel) %>%
  mutate(zscore_perROI_perrun_persubj = (beta - mean(beta)) / sd(beta)) %>%
  ungroup() %>% 
  as.data.frame()
```


```{r}
MVPA_vectors_runs12_zscored <-
  MVPA_vectors_runs12_zscored0 %>%
  select(-beta) %>%
  spread(voxel, zscore_perROI_perrun_persubj)

# saveRDS(
#   MVPA_vectors_runs12_zscored,
#   here(
#     "outputs",
#     "multivariate_data",
#     "study2_multivariate_data_runs12_top100_zscored.Rds"
#   )
# )
```


```{r}

MVPA_data <-
  readRDS(
    here(
      "outputs",
      "multivariate_data",
      "study2_multivariate_data_runs12_top100.Rds"
    )
  )


MVPA_data_zscored <-
  readRDS(here(
    "outputs",
    "multivariate_data",
    "study2_multivariate_data_runs12_top100_zscored.Rds"
  ))


subjects <- levels(as.factor(MVPA_data$subj))
ROIs <- levels(as.factor(MVPA_data$ROI))


distance_metrics <- c("Pearson", "Euclidean")
labels <- c('pearson', 'euclidean')

# for each distance metric...
for (distance_metricI in 1:length(distance_metrics)) {

  
  distance_metric <- distance_metrics[distance_metricI]
  distance_metric_lower <- tolower(distance_metric)
  distance_metric_label <- labels[distance_metricI]

  # initialize final output
  similarity_output_matrix <- data.frame()

  # GET EMPIRICAL RDMS
  # for each subject and ROI...
  for (i in 1:length(subjects)) {
    for (k in 1:length(ROIs)) {

      # get data specific to subject and ROI
      subjROIdata <- MVPA_data %>%
        filter(subj == subjects[i],
               # event != "fam", # only take test events
               ROI == ROIs[k]) %>%
        select(extracted_run_number, domain, event, cond:V100) %>%
        unite(
          col = "cond_run",
          c("cond", "extracted_run_number"),
          sep = "_",
          remove = FALSE
        ) %>%
        select(-domain, -event, -extracted_run_number, -cond) %>%
        as.data.frame()

      row.names(subjROIdata) <- subjROIdata$cond_run

      subjROIdata.final <- subjROIdata %>% select(-cond_run)

      # pearson
      if (distance_metric == 'Pearson') {
        distmatrix <- 1 - cor(t(subjROIdata.final)) # convert similarity into a distance (1-r)
      }

      # euclidian
      if (distance_metric == 'Euclidean') {
        distmatrix <- dist(subjROIdata.final, 
                           method = "euclidean", 
                           diag = TRUE, 
                           upper = TRUE) %>% 
          as.matrix()
      }

      cur.subset <- distmatrix %>%
          as.data.frame() %>%
          # select(-subj, -ROI, -dist) %>%
          flattenCorrMatrix(.) %>%
          unite(pair, c(row,column), sep = "-vs-", remove = FALSE) %>%
          rename(dist = cor) %>%
          mutate(subj = subjects[i],
                 ROI = ROIs[k],
                 distance_metric = distance_metric_label) %>%
        
          # only take distances across runs
          mutate(same_run = case_when((str_detect(row, "_run1") & str_detect(column, "_run1")) ~ 1,
                                      (str_detect(row, "_run2") & str_detect(column, "_run2")) ~ 1,
                                      TRUE ~ 0
                                      ))
          # filter(same_run == 0) 
        similarity_output_matrix <- rbind(similarity_output_matrix, cur.subset)
      }
  }

# save out RDM.empirical.model.distZ.all_iters
  fname =  paste0(here("outputs/multivariate_data/"),
                  "MVPA_subject_RDMs_runs12_top100_",
                  distance_metric,
                  # ".Rds",
                  "_z.Rds",
                  sep = "")
  
    saveRDS(similarity_output_matrix, fname)
}

```

```{r}
MVPA_subject_RDMs_Euclidean_zscored <- readRDS(here("outputs/multivariate_data/MVPA_subject_RDMs_runs12_top100_Euclidean_z.Rds")) %>%
  mutate(distance_metric = "euclidean_zscored")

MVPA_subject_RDMs_Euclidean <- readRDS(here("outputs/multivariate_data/MVPA_subject_RDMs_runs12_top100_Euclidean.Rds")) %>%
  mutate(distance_metric = "euclidean")

MVPA_subject_RDMs_Pearson_zscored <- readRDS(here("outputs/multivariate_data/MVPA_subject_RDMs_runs12_top100_Pearson_z.Rds")) %>%
  mutate(distance_metric = "pearson_zscored")

MVPA_subject_RDMs_Pearson <- readRDS(here("outputs/multivariate_data/MVPA_subject_RDMs_runs12_top100_Pearson.Rds")) %>%
  mutate(distance_metric = "pearson")

MVPA_combined_RDMs <- rbind(
  MVPA_subject_RDMs_Euclidean_zscored,
  MVPA_subject_RDMs_Euclidean,
  MVPA_subject_RDMs_Pearson_zscored,
  MVPA_subject_RDMs_Pearson
) %>%
  separate(row,
           c("row_domain", "row_event", "row_run"),
           remove = FALSE) %>%
  separate(column,
           c("column_domain", "column_event", "column_run"),
           remove = FALSE) %>%
  
  # for this col, we want to label all events matching on domain as within, and not matching as between
  mutate(pair_label_domain =
           case_when((row_domain == column_domain) ~ "within",
                     TRUE ~ "between")) %>%
  
  # for this col, we want to label all events matching on event as within, not matching as between
  mutate(pair_label_event =
           case_when((row_event == column_event) ~ "within",
                     TRUE ~ "between")) %>%
  
  # same as matching by event, except we drop labels from the opposite domain
  mutate(
    pair_label_event_within_domain =
      case_when(
        # same domain and event
        (pair_label_domain == "within" &
           pair_label_event == "within") ~ "within",
        
        # same domain, different event
        (pair_label_domain == "within" &
           pair_label_event == "between") ~ "between",
        
        # drop opposite domain
        pair_label_domain == "between" ~ "drop_domain"
      )
  ) %>% 
  
  # domains within events (rather than events within domains)
  mutate(pair_label_domain_within_event =
           case_when(
             # same domain, same event
             (pair_label_domain == "within" &
                pair_label_event == "within") ~ "within",
             
             # opposite domain
             (pair_label_domain == "between" &
                pair_label_event == "within")  ~ "between",
             
             # drop opposite event
             (pair_label_event == "between") ~ "drop_event"
           )) %>%
  
  # events ACROSS domains (true "domain general" metric)
  mutate(pair_label_event_across_domains =
           case_when(
             # same event, opposite domain
             (pair_label_domain == "between" &
                pair_label_event == "within") ~ "within",
             
             # different event, different domain
             (pair_label_domain == "between" &
                pair_label_event == "between")  ~ "between",
             
             # drop opposite domain
             TRUE ~ "drop_domain"
           )) %>%
  
  # domains ACROSS events (true "event general" metric)
  mutate(pair_label_domains_across_events =
           case_when(
             # same domain, opposite event
             (pair_label_event == "between" &
                pair_label_domain == "within") ~ "within",
             
             # different domain, different event
             (pair_label_event == "between" &
                pair_label_domain == "between")  ~ "between",
             
             # drop opposite event
             TRUE ~ "drop_event"
           )) %>%
  
  mutate(same_domain = case_when(
    (
      pair_label_event_within_domain != "drop_domain"
      & pair_label_domain == "within"
    ) ~ row_domain
  )) %>%
  
  mutate(same_event = case_when(
    (
      pair_label_domain_within_event != "drop_event" &
        pair_label_event == "within"
    ) ~ row_event
  )) %>%
  
  relocate(pair, .after = last_col())

```



```{r, include = TRUE}
ROIs <- levels(as.factor(MVPA_combined_RDMs$ROI))
subjects <- levels(as.factor(MVPA_combined_RDMs$subj))

RDMs <-
  c(
    "pair_label_domain",
    "pair_label_event",
    "pair_label_event_within_domain",
    "pair_label_domain_within_event",
    "pair_label_event_across_domains",
    "pair_label_domains_across_events"
  )

MVPA_combined_RDMs_out0 <- MVPA_combined_RDMs %>%
  rename(ROI_name_final = ROI) %>%
  mutate_at(RDMs, as.factor) %>%
  mutate_at(RDMs, relev)
```

```{r}
region_info <- read.csv(here("input_data/manyregions_info.csv")) %>%
  mutate(ROI_name = paste0(ROI_name, "_MNI152Lin2009cAsym_res-02", sep = ""))
```


```{r, include = TRUE}
MVPA_combined_RDMs_out <-
  full_join(MVPA_combined_RDMs_out0, region_info) %>%
  mutate(
    ROI_side = case_when(str_detect(ROI_name_final, "bilateral") ~ "bilateral",
                      TRUE ~ str_sub(ROI_name_final,-1,-1)),
    ROI_simplename = str_split(ROI_name_final, "_")[[1]][1]
  )

MVPA_combined_RDMs_out$ROI_category <-
  factor(
    MVPA_combined_RDMs_out$ROI_category,
    levels = c("physics", "psychology", "MD", "early_visual")
  )

MVPA_combined_RDMs_out$distance_metric <-
  factor(
    MVPA_combined_RDMs_out$distance_metric,
    levels = c("euclidean", "euclidean_zscored", "pearson", "pearson_zscored")
  )

# saveRDS(MVPA_combined_RDMs_out, here("outputs", "multivariate_data", "study2_MVPA_processed_distances_top100_runs12.Rds"))
```


### Secondary analysis
Here we want to take data from runs 1-4, all events, and compute distances over them.

```{r}
MVPA_vectors_allevents_allfolds <-
  readRDS(
    here("outputs/multivariate_data/study2_multivariate_data_acrossfolds_top100.Rds")
  )

# are the number of observations per region per subject expected? n_observations should be 9 (2 runs, 3 domains, 3 events) across 4 runs, 3 per run
MVPA_vectors_allevents_allfolds %>%
  group_by(subj, ROI, extracted_run_number) %>%
  summarise(n_observations = n()) 

table(MVPA_vectors_allevents_allfolds$event)
table(MVPA_vectors_allevents_allfolds$length_cope_actual)
```

```{r}
# z score vectors within voxels within runs within ROIs within subjects
MVPA_vectors_allevents_allfolds_zscored0 <- MVPA_vectors_allevents_allfolds %>%
  gather(voxel, beta, V1:V100) %>%
  group_by(subj, ROI, extracted_run_number, voxel) %>%
  mutate(zscore_perROI_perrun_persubj = (beta - mean(beta)) / sd(beta)) %>%
  ungroup() %>% 
  as.data.frame()
```


```{r}
MVPA_vectors_allevents_allfolds_zscored <-
  MVPA_vectors_allevents_allfolds_zscored0 %>%
  select(-beta) %>%
  spread(voxel, zscore_perROI_perrun_persubj)

saveRDS(
  MVPA_vectors_allevents_allfolds_zscored,
  here(
    "outputs",
    "multivariate_data",
    "study2_multivariate_data_allevents_allfolds_zscored.Rds"
  )
)
```


```{r}

MVPA_data <-
  readRDS(
    here(
      "outputs",
      "multivariate_data",
      "study2_multivariate_data_acrossfolds_top100.Rds"
    )
  )


MVPA_data_zscored <-
  readRDS(
    here(
      "outputs",
      "multivariate_data",
      "study2_multivariate_data_allevents_allfolds_zscored.Rds"
    )
  )


subjects <- levels(as.factor(MVPA_data$subj))
ROIs <- levels(as.factor(MVPA_data$ROI))


distance_metrics <- c("Pearson", "Euclidean")
labels <- c('pearson', 'euclidean')

# for each distance metric...
for (distance_metricI in 1:length(distance_metrics)) {

  
  distance_metric <- distance_metrics[distance_metricI]
  distance_metric_lower <- tolower(distance_metric)
  distance_metric_label <- labels[distance_metricI]

  # initialize final output
  similarity_output_matrix <- data.frame()

  # GET EMPIRICAL RDMS
  # for each subject and ROI...
  for (i in 1:length(subjects)) {
    for (k in 1:length(ROIs)) {

      # get data specific to subject and ROI
      subjROIdata <- MVPA_data %>%
        filter(subj == subjects[i],
               # event != "fam", # only take test events
               ROI == ROIs[k]) %>%
        select(extracted_run_number, domain, event, cond:V100) %>%
        unite(
          col = "cond_run",
          c("cond", "extracted_run_number"),
          sep = "_",
          remove = FALSE
        ) %>%
        select(-domain, -event, -extracted_run_number, -cond) %>%
        as.data.frame()

      row.names(subjROIdata) <- subjROIdata$cond_run

      subjROIdata.final <- subjROIdata %>% select(-cond_run)

      # pearson
      if (distance_metric == 'Pearson') {
        distmatrix <- 1 - cor(t(subjROIdata.final)) # convert similarity into a distance (1-r)
      }

      # euclidian
      if (distance_metric == 'Euclidean') {
        distmatrix <- dist(subjROIdata.final, 
                           method = "euclidean", 
                           diag = TRUE, 
                           upper = TRUE) %>% 
          as.matrix()
      }

      cur.subset <- distmatrix %>%
          as.data.frame() %>%
          # select(-subj, -ROI, -dist) %>%
          flattenCorrMatrix(.) %>%
          filter(str_detect(row, "exclude") | str_detect(column, "exclude")) %>%
          filter(str_sub(row, -1) == str_sub(column, -1)) %>%# only take rows that compare a run against correponding fold
          unite(pair, c(row,column), sep = "-vs-", remove = FALSE) %>%
          rename(dist = cor) %>%
          mutate(subj = subjects[i],
                 ROI = ROIs[k],
                 distance_metric = distance_metric_label)
        
        similarity_output_matrix <- rbind(similarity_output_matrix, cur.subset)
      }
  }

# save out RDM.empirical.model.distZ.all_iters
  fname =  paste0(here("outputs/multivariate_data/"),
                  "MVPA_subject_RDMs_allevents_allfolds_top100_",
                  distance_metric,
                  # ".Rds",
                  "_z.Rds",
                  sep = "")
  
    saveRDS(similarity_output_matrix, fname)
}

```

```{r}
MVPA_subject_RDMs_Euclidean_zscored <- readRDS(here("outputs/multivariate_data/MVPA_subject_RDMs_allevents_allfolds_top100_Euclidean_z.Rds")) %>%
  mutate(distance_metric = "euclidean_zscored")

MVPA_subject_RDMs_Euclidean <- readRDS(here("outputs/multivariate_data/MVPA_subject_RDMs_allevents_allfolds_top100_Euclidean.Rds")) %>%
  mutate(distance_metric = "euclidean")

MVPA_subject_RDMs_Pearson_zscored <- readRDS(here("outputs/multivariate_data/MVPA_subject_RDMs_allevents_allfolds_top100_Pearson_z.Rds")) %>%
  mutate(distance_metric = "pearson_zscored")

MVPA_subject_RDMs_Pearson <- readRDS(here("outputs/multivariate_data/MVPA_subject_RDMs_allevents_allfolds_top100_Pearson.Rds")) %>%
  mutate(distance_metric = "pearson")

MVPA_combined_RDMs <- rbind(
  MVPA_subject_RDMs_Euclidean_zscored,
  MVPA_subject_RDMs_Euclidean,
  MVPA_subject_RDMs_Pearson_zscored,
  MVPA_subject_RDMs_Pearson
) %>%
  separate(row,
           c("row_domain", "row_event", "row_run"),
           remove = FALSE) %>%
  separate(column,
           c("column_domain", "column_event", "column_run"),
           remove = FALSE) %>%
  
  filter(row_run != column_run) %>%
  # for this col, we want to label all events matching on domain as within, and not matching as between
  mutate(pair_label_domain =
           case_when((row_domain == column_domain) ~ "within",
                     TRUE ~ "between")) %>%
  
  # for this col, we want to label all events matching on event as within, not matching as between
  mutate(pair_label_event =
           case_when((row_event == column_event) ~ "within",
                     TRUE ~ "between")) %>%
  
  # same as matching by event, except we drop labels from the opposite domain
  mutate(
    pair_label_event_within_domain =
      case_when(
        # same domain and event
        (pair_label_domain == "within" &
           pair_label_event == "within") ~ "within",
        
        # same domain, different event
        (pair_label_domain == "within" &
           pair_label_event == "between") ~ "between",
        
        # drop opposite domain
        pair_label_domain == "between" ~ "drop_domain"
      )
  ) %>% 
  
  # domains within events (rather than events within domains)
  mutate(pair_label_domain_within_event =
           case_when(
             # same domain, same event
             (pair_label_domain == "within" &
                pair_label_event == "within") ~ "within",
             
             # opposite domain
             (pair_label_domain == "between" &
                pair_label_event == "within")  ~ "between",
             
             # drop opposite event
             (pair_label_event == "between") ~ "drop_event"
           )) %>%
  
  # events ACROSS domains (true "domain general" metric)
  mutate(pair_label_event_across_domains =
           case_when(
             # same event, opposite domain
             (pair_label_domain == "between" &
                pair_label_event == "within") ~ "within",
             
             # different event, different domain
             (pair_label_domain == "between" &
                pair_label_event == "between")  ~ "between",
             
             # drop opposite domain
             TRUE ~ "drop_domain"
           )) %>%
  
  # domains ACROSS events (true "event general" metric)
  mutate(pair_label_domains_across_events =
           case_when(
             # same domain, opposite event
             (pair_label_event == "between" &
                pair_label_domain == "within") ~ "within",
             
             # different domain, different event
             (pair_label_event == "between" &
                pair_label_domain == "between")  ~ "between",
             
             # drop opposite event
             TRUE ~ "drop_event"
           )) %>%
  
  mutate(same_domain = case_when(
    (
      pair_label_event_within_domain != "drop_domain"
      & pair_label_domain == "within"
    ) ~ row_domain
  )) %>%
  
  mutate(same_event = case_when(
    (
      pair_label_domain_within_event != "drop_event" &
        pair_label_event == "within"
    ) ~ row_event
  )) %>%
  
  relocate(pair, .after = last_col())

```

```{r}
MVPA_combined_RDMs %>%
  filter(subj == "sub-SAXNES2s001",
         ROI == "antParietal_bilateral") %>%
  View()
```



```{r, include = TRUE}
ROIs <- levels(as.factor(MVPA_combined_RDMs$ROI))
subjects <- levels(as.factor(MVPA_combined_RDMs$subj))

RDMs <-
  c(
    "pair_label_domain",
    "pair_label_event",
    "pair_label_event_within_domain",
    "pair_label_domain_within_event",
    "pair_label_event_across_domains",
    "pair_label_domains_across_events"
  )

MVPA_combined_RDMs_out0 <- MVPA_combined_RDMs %>%
  rename(ROI_name_final = ROI) %>%
  mutate_at(RDMs, as.factor) %>%
  mutate_at(RDMs, relev)
```

```{r}
region_info <- read.csv(here("input_data/manyregions_info.csv")) %>%
  mutate(ROI_name = paste0(ROI_name, "_MNI152Lin2009cAsym_res-02", sep = ""))
```


```{r, include = TRUE}
MVPA_combined_RDMs_out <-
  full_join(MVPA_combined_RDMs_out0, region_info) %>%
  mutate(
    ROI_side = case_when(str_detect(ROI_name_final, "bilateral") ~ "bilateral",
                      TRUE ~ str_sub(ROI_name_final,-1,-1)),
    ROI_simplename = str_split(ROI_name_final, "_")[[1]][1]
  )

MVPA_combined_RDMs_out$ROI_category <-
  factor(
    MVPA_combined_RDMs_out$ROI_category,
    levels = c("physics", "psychology", "MD", "early_visual")
  )

MVPA_combined_RDMs_out$distance_metric <-
  factor(
    MVPA_combined_RDMs_out$distance_metric,
    levels = c("euclidean", "euclidean_zscored", "pearson", "pearson_zscored")
  )

saveRDS(MVPA_combined_RDMs_out, here("outputs", "multivariate_data", "study2_MVPA_processed_distances_top100_allfolds.Rds"))
```


## Compute the noise ceiling per ROI

```{r}
study2_distances_runs12 <- readRDS(here("outputs/multivariate_data/study2_MVPA_processed_distances_top100_runs12.Rds"))
str(study2_distances_runs12)

subjects <- unique(study2_distances_runs12$subj)
ROIs <- unique(study2_distances_runs12$ROI_name_final)
```

```{r}


study2_distances_runs12_acrossruns <- study2_distances_runs12 %>%
      filter(same_run == 1,
             distance_metric == "euclidean",
             str_detect(pair, "both", negate = TRUE),
             str_detect(pair, "fam", negate = TRUE))

noise_ceiling_acrossrun <- data.frame()

euclidean_dist <- function(x, y) sqrt(sum((x - y)^2))


for (ROI in ROIs) {
  for (subject in subjects) {
    
    # current subject
    cur_subject <- study2_distances_runs12_acrossruns %>%
      filter(subj == subject,
             ROI_name_final == ROI) %>%
      select(pair, dist) %>%
      arrange(pair)

    # everyone including current subject
    everyone <- study2_distances_runs12_acrossruns %>%
      filter(ROI_name_final == ROI) %>%
      group_by(pair) %>%
      summarise(dist_avg_everyone = mean(dist)) %>%
      arrange(pair)
    
    # everyone except the current subject    
    everyone_else <- study2_distances_runs12_acrossruns %>%
      filter(subj != subject,
             ROI_name_final == ROI) %>%
      group_by(pair) %>%
      summarise(dist_avg_everyonebutX = mean(dist)) %>%
      arrange(pair)
    
    # compute lower bound noise ceiling, excluding cur subj
    euclidean_cur_iteration_lowerbound <-
      euclidean_dist(cur_subject$dist,
          everyone_else$dist_avg_everyonebutX)
    
    kendall_cur_iteration_lowerbound <-
      cor(cur_subject$dist,
          everyone_else$dist_avg_everyonebutX,
          method = "kendall")
    
    # compute upper bound noise ceiling, including cur subj
    euclidean_cur_iteration_upperbound <-
      euclidean_dist(cur_subject$dist,
          everyone$dist_avg_everyone)
    
    kendall_cur_iteration_upperbound <-
      cor(cur_subject$dist,
          everyone$dist_avg_everyone,
          method = "kendall")
    
    cur_row <-
      cbind(subject, 
            ROI, 
            euclidean_cur_iteration_lowerbound,
            euclidean_cur_iteration_upperbound,
            kendall_cur_iteration_lowerbound,
            kendall_cur_iteration_upperbound
      )
    
    noise_ceiling_acrossrun <- rbind(noise_ceiling_acrossrun, cur_row)
  }
}

        

noise_ceiling_average_acrossrun <- noise_ceiling_acrossrun %>%
  mutate_at(c("kendall_cur_iteration_lowerbound",
              "kendall_cur_iteration_upperbound",
              "euclidean_cur_iteration_lowerbound",
              "euclidean_cur_iteration_upperbound"), as.numeric) %>%
  group_by(ROI) %>%
  summarise(kendall_ceiling_lowerbound_avg = mean(kendall_cur_iteration_lowerbound),
            kendall_ceiling_upperbound_avg = mean(kendall_cur_iteration_upperbound),
            euclidean_ceiling_lowerbound_avg = mean(euclidean_cur_iteration_lowerbound),
            euclidean_ceiling_upperbound_avg = mean(euclidean_cur_iteration_upperbound))

noise_ceiling_average_acrossrun_long <- noise_ceiling_average_acrossrun %>%
  gather(metric, value, kendall_ceiling_lowerbound_avg:euclidean_ceiling_upperbound_avg) %>%
  mutate(cor_metric = case_when(str_detect(metric, "kendall") ~ "kendall",
                                str_detect(metric, "euclidean") ~ "euclidean"
         ),
         upper_lower = case_when(str_detect(metric, "lower") ~ "lower",
                                 str_detect(metric, "upper") ~ "upper"))
```

```{r}
noise_ceiling_average_acrossrun_long_lower_upper <- noise_ceiling_average_acrossrun_long %>%
  pivot_wider(id_cols = c("ROI", "cor_metric"), names_from = c("upper_lower")) %>%
  group_by(ROI, cor_metric) %>%
  mutate(average = mean(c(lower, upper))) %>%
  ungroup() %>%
    rename(noiseceiling_lower = lower,
         noiseceiling_upper = upper,
         noiseceiling_average = average,
         ROI_name_final = ROI)


noise_ceiling_average_acrossrun_long_lower_upper_final <- left_join(noise_ceiling_average_acrossrun_long_lower_upper, region_info)
```

```{r}
# write_rds(noise_ceiling_average_acrossrun_long_lower_upper_final, path = here("outputs/multivariate_results/MVPA_noiseceiling_withinrun_perROI.Rds"))
```

```{r}
MVPA_noiseceiling_withinrun <- read_rds(here("outputs/multivariate_results/study2_MVPA_noiseceiling_withinrun_perROI.Rds")) %>%
  mutate(ceiling = "within_run")

MVPA_noiseceiling_acrossrun <- read_rds(here("outputs/multivariate_results/study2_MVPA_noiseceiling_perROI.Rds")) %>%
  mutate(ceiling = "across_run")

MVPA_noiseceiling <- rbind(MVPA_noiseceiling_acrossrun, MVPA_noiseceiling_withinrun) %>%
  filter(focal_region == 1,
         cor_metric == "euclidean")
```

```{r}

ggplot(MVPA_noiseceiling, aes(ceiling, noiseceiling_upper)) +
  geom_boxplot() +
  geom_point() +
  geom_line(aes(group = ROI_name_final, alpha = .4)) +
  xlab("Ceiling Computed Within or Across Runs") +
  ylab("Upper bound noise ceiling (Euclidean)") +
  geom_text_repel(aes(label = ROI_name_final)) +
  theme(legend.position = "none")
```

# Behavioral ratings

## Psychology

```{r}
efficiency_ratings <- py_load_object(here("input_data/behavioral_ratings/agent_behavioral_ratings_scripts/types_raw_mean_action_efficiency.pik"))

efficiency_ratings_processed <- unlist(efficiency_ratings, recursive = FALSE) 
efficiency_ratings_processed_df <- as.data.frame(efficiency_ratings_processed) 

efficiency_ratings_final <- efficiency_ratings_processed_df %>%
  pivot_longer(cols = expected.type_1.Set_1.v1:surprising.type_5.Set_9.v2,
               names_to = "trial",
               values_to = "rating") %>%
  separate(trial, sep = "[.]", into = c("trial_type", "scenario", "set", "version"), remove = FALSE) %>%
  filter(scenario == "type_1" | scenario == "type_2" | scenario == "type_5") %>%
  mutate(task = case_when((scenario == "type_1" | scenario == "type_2") ~"efficiency",
                          scenario == "type_5" ~ "agent_solidity")) %>%
  mutate(trial_type = case_when(str_detect(trial, "expected") ~ "expected",
                              TRUE ~ "unexpected")) %>%
  unite("trial", c(scenario, set, version)) %>%
  select("trial", "rating", "trial_type", "task") %>%
  mutate(trial_id = trial)

```

```{r}
goal_ratings <- py_load_object(here("input_data/behavioral_ratings/agent_behavioral_ratings_scripts/raw_mean_goal_preferences.pik"))

goal_ratings_processed <- unlist(goal_ratings, recursive = FALSE) 
goal_ratings_processed_df <- as.data.frame(goal_ratings_processed) 

goal_ratings_final <- goal_ratings_processed_df %>%
  pivot_longer(cols = expected1:surprising40,
               names_to = "trial",
               values_to = "rating") %>%
  mutate(trial_type = case_when(str_detect(trial, "expected") ~ "expected",
                                TRUE ~ "unexpected")) %>%
  mutate(task = "goal") %>%
  mutate(trial_id = as.character(parse_number(trial)))
```

```{r}
infer_constraints_ratings <- py_load_object(here("input_data/behavioral_ratings/agent_behavioral_ratings_scripts/raw_mean_unobserved_constraints.pik"))

infer_constraints_ratings_processed <- unlist(infer_constraints_ratings, recursive = FALSE) 
infer_constraints_ratings_processed_df <- as.data.frame(infer_constraints_ratings_processed) 

infer_constraints_ratings_final <- infer_constraints_ratings_processed_df %>%
  pivot_longer(cols = expected1:surprising20,
               names_to = "trial",
               values_to = "rating") %>%
  mutate(trial_type = case_when(str_detect(trial, "expected") ~ "expected",
                                TRUE ~ "unexpected")) %>%
  mutate(task = "infer_constraint") %>%
  mutate(trial_id = as.character(parse_number(trial)))
```

```{r}
agent_ratings <- rbind(efficiency_ratings_final,
                       goal_ratings_final,
                       infer_constraints_ratings_final) 

efficiency_model <- lmer(data = agent_ratings %>% filter(task == "efficiency"),
                         formula = rating ~ trial_type + (1|trial_id))

efficiency_d <- lme.dscore(efficiency_model, agent_ratings %>% filter(task == "efficiency"), "lme4") %>%
        select(d) %>%
  as.numeric()

goal_model <- lmer(data = agent_ratings %>% filter(task == "goal"),
                         formula = rating ~ trial_type + (1|trial_id))

goal_d <- lme.dscore(goal_model, agent_ratings %>% filter(task == "goal"), "lme4") %>%
        select(d) %>%
  as.numeric()

infer_constraint_model <- lmer(data = agent_ratings %>% filter(task == "infer_constraint"),
                         formula = rating ~ trial_type + (1|trial_id))

infer_constraint_d <- lme.dscore(infer_constraint_model, agent_ratings %>% filter(task == "infer_constraint"), "lme4") %>%
        select(d) %>%
  as.numeric()

agent_solidity_model <- lmer(data = agent_ratings %>% filter(task == "agent_solidity"),
                         formula = rating ~ trial_type + (1|trial_id))

agent_solidity_d <- lme.dscore(infer_constraint_model, agent_ratings %>% filter(task == "infer_constraint"), "lme4") %>%
        select(d) %>%
  as.numeric()

```

```{r}
psych_action_model <- lmer(data = agent_ratings %>% filter(task == "goal" | task == "efficiency"),
                         formula = rating ~ trial_type + (1|task))

summary(psych_action_model)

psych_action_model_d <- lme.dscore(psych_action_model, agent_ratings %>% filter(task == "goal" | task == "efficiency"), "lme4") %>%
        select(d) %>%
  as.numeric()
```

```{r}
psych_env_model <- lmer(data = agent_ratings %>% filter(task == "agent_solidity" | task == "infer_constraint"),
                         formula = rating ~ trial_type + (1|task))

summary(psych_env_model)

psych_env_model_d <- lme.dscore(psych_env_model, agent_ratings  %>% filter(task == "agent_solidity" | task == "infer_constraint"), "lme4") %>%
        select(d) %>%
  as.numeric()
```

## Physics
```{r}
conditions <- c("create", "block", "long-overturn", "invisible-discontinuous")

physics_ratings_init <- read.csv(here("input_data/behavioral_ratings/adept_physics_surprise_ratings.csv")) %>%
  filter(SurpriseCondition %in% conditions) %>%
  mutate(trial_type = case_when(str_detect(SurpriseType, "control") ~ "expected",
                                str_detect(SurpriseType, "surprise") ~ "unexpected")) %>%
  mutate(task = case_when(SurpriseCondition == "create" ~ "permanence",
                          SurpriseCondition == "invisible-discontinuous" ~ "permanence",
                          SurpriseCondition == "block" ~ "solidity",
                          SurpriseCondition == "long-overturn" ~ "solidity")) %>%
  filter(SurpriseCondition != "invisible-discontinuous") # the exp / unexp mappings are backwards for this scenario; we'll fix that below

physics_ratings_invisible_discontinuous <- read.csv(here("input_data/behavioral_ratings/adept_physics_surprise_ratings.csv")) %>%
  filter(SurpriseCondition == "invisible-discontinuous") %>%
  mutate(trial_type = case_when(str_detect(SurpriseType, "surprise") ~ "expected",
                                str_detect(SurpriseType, "control") ~ "unexpected")) %>%
  mutate(task = "permanence")

physics_ratings <- rbind(physics_ratings_init, physics_ratings_invisible_discontinuous) %>%
  group_by(task, SurpriseCondition, ClassName,  trial_type) %>%
  summarise(rating = mean(SurpriseRating)) 
```

```{r}
physics_ratings %>%
  group_by(SurpriseCondition) %>%
  summarise(n = n())

length(unique(physics_ratings$UID))
```


```{r}
permanence_model <- lmer(data = physics_ratings %>% filter(task == "permanence"),
                         formula = rating ~ trial_type + (1|ClassName))

summary(permanence_model)

permanence_model_d <- lme.dscore(permanence_model, physics_ratings %>% filter(task == "permanence"), "lme4") %>%
        select(d) %>%
  as.numeric()
```


```{r}
solidity_model <- lmer(data = physics_ratings %>% filter(task == "solidity"),
                         formula = rating ~ trial_type + (1|ClassName))

summary(solidity_model)

solidity_model_d <- lme.dscore(solidity_model, physics_ratings %>% filter(task == "solidity"), "lme4") %>%
        select(d) %>%
  as.numeric()
```


```{r}
physics_model <- lmer(data = physics_ratings,
                         formula = rating ~ trial_type + (1|task))

summary(physics_model)

physics_model_d <- lme.dscore(physics_model, physics_ratings, "lme4") %>%
        select(d) %>%
  as.numeric()
```

```{r}
task_domain <-
  c("efficiency",
    "goal",
    "infer_constraint",
    "agent_solidity",
    "permanence",
    "solidity",
    "psych_action",
    "psych_environment",
    "physics")

cohens_d <-
  c(efficiency_d,
    goal_d,
    infer_constraint_d,
    agent_solidity_d,
    permanence_model_d,
    solidity_model_d,
    psych_action_model_d,
    psych_env_model_d,
    physics_model_d)

effect_sizes <- cbind(task_domain, cohens_d) %>%
  as.data.frame()
```

## Plots

```{r}
theme_set(theme_cowplot(15))
```

```{r}
effect_sizes_per_domain <- effect_sizes %>% slice(7:9)
effect_sizes_per_task <- effect_sizes %>% slice(1:6)
```

```{r}
psych_action_plot_task <- ggplot(data = agent_ratings %>% filter(task == "goal" | task == "efficiency"), aes(trial_type, rating)) +
  geom_boxplot() +
  geom_point() +
  coord_cartesian(ylim = c(0,100)) +
  # geom_text(data = effect_sizes_per_task %>% filter(task == "goal" | task == "efficiency")) +
  facet_wrap(~task, nrow = 1) +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
```

```{r}
psych_env_plot_task <- ggplot(data = agent_ratings %>% filter(task == "infer_constraint" | task == "agent_solidity"), aes(trial_type, rating)) +
  geom_boxplot() +
  geom_point() +
  coord_cartesian(ylim = c(0,100)) +
  facet_wrap(~task, nrow = 1) +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
```

```{r}
psych_action_plot_domain <- ggplot(data = agent_ratings %>% filter(task == "goal" | task == "efficiency"), aes(trial_type, rating)) +
  geom_boxplot() +
  geom_point() +
  coord_cartesian(ylim = c(0,100)) +
  ggtitle("AGENT (PSYCH ACTION)")
```

```{r}
psych_env_plot_domain <- ggplot(data = agent_ratings %>% filter(task == "infer_constraint" | task == "agent_solidity"), aes(trial_type, rating)) +
  geom_boxplot() +
  geom_point() +
  coord_cartesian(ylim = c(0,100)) +
  ggtitle("AGENT (PSYCH ENV)")
```


```{r}
phys_plot_task <- ggplot(data = physics_ratings, aes(trial_type, rating)) +
  geom_boxplot() +
  geom_point() +
  coord_cartesian(ylim = c(0,100)) +
  facet_wrap(~task, nrow = 1) +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
```

```{r}
phys_plot_domain <- ggplot(data = physics_ratings, aes(trial_type, rating)) +
  geom_boxplot() +
  geom_point() +
  coord_cartesian(ylim = c(0,100)) +
  ggtitle("ADEPT (PHYSICS)")
```

```{r}
domain_plots <- ( psych_action_plot_domain | psych_env_plot_domain | phys_plot_domain | gridExtra::tableGrob(effect_sizes_per_domain))
task_plots <- (psych_action_plot_task | psych_env_plot_task | phys_plot_task | gridExtra::tableGrob(effect_sizes_per_task)) 

(domain_plots / task_plots) + plot_annotation(tag_levels = 'A')
```
